Add betty 100-prime_factor.c

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-unsigned short int n = 612852475143, a = (int)sqrt(n);
+unsigned long int n = 612852475143, a = (int)sqrt(n);
 while (1)
 {
 if (n % a == 0)
